bugfix for first run of gui
