+++ aetd_modules/__init__.py
@@ -1,19 +1,27 @@
-from .data_containers import (
+from .containers import (
     AnnotationsContainer,
     DirectionBox,
+    Driveable,
+    Impassable,
+    Passable,
+    Path,
     PathsBox,
     RoadObjectsBox,
     RoadSegmentsBox,
+    Sign,
     SpeedBox,
+    TrafficLight,
+    Vehicle,
 )
 from .direction import DirectionExtractor
 from .draw import Draw
-from .paths import Path, PathExtractor, PathPlanner
+from .paths import PathExtractor, PathPlanner
 from .pipeline import Pipeline
 from .preprocessor import Preprocessor
-from .road_objects import RoadObjectExtractor, Sign, TrafficLight, Vehicle
-from .road_segmentations import Driveable, Impassable, Passable, RoadSegmentsExtractor
+from .road_objects import RoadObjectExtractor
+from .road_segmentations import RoadSegmentsExtractor
 from .speed import SpeedDataExtractor
+from .types import Float, Img, ImgT, LImg, LImgT
 
 __all__: list[str] = [
     "DirectionExtractor",
@@ -29,8 +37,8 @@ __all__: list[str] = [
     "PathPlanner",
     "PathExtractor",
     "Path",
-    "Preprocessor",
     "RoadSegmentsBox",
+    "Preprocessor",
     "RoadObjectsBox",
     "SpeedBox",
     "DirectionBox",
@@ -38,4 +46,9 @@ __all__: list[str] = [
     "AnnotationsContainer",
     "Draw",
     "Pipeline",
+    "Float",
+    "Img",
+    "ImgT",
+    "LImg",
+    "LImgT",
 ]
