+++ debug/__init__.py
@@ -1,3 +1,25 @@
-from layouts import MainWindow
+from .components import (
+    DirectionsTab,
+    FullAnnotationTab,
+    ImageViewer,
+    ModuleTabBar,
+    ModulTab,
+    PathTab,
+    RoadObjectTab,
+    SegmentorTab,
+    SpeedTab,
+)
+from .layouts import MainWindow
 
-__all__: list[str] = ["MainWindow"]
+__all__: list[str] = [
+    "MainWindow",
+    "ImageViewer",
+    "ModuleTabBar",
+    "ModulTab",
+    "DirectionsTab",
+    "SpeedTab",
+    "RoadObjectTab",
+    "SegmentorTab",
+    "PathTab",
+    "FullAnnotationTab",
+]
