+++ debug/components/tab_bar.py
@@ -6,39 +6,39 @@ from .tabs import DirectionsTab, FullAnnotationTab, PathTab, RoadObjectTab, Segm
 
 class ModuleTabBar(QWidget):
     def __init__(self, parent: QWidget | None = None) -> None:
-        super().__init__(parent=parent)
+        super().__init__(parent)
 
         # layout to hold both the tab bar and the button
-        layout = QHBoxLayout(parent=self)
-        layout.setContentsMargins(left=0, top=0, right=0, bottom=0)
+        layout = QHBoxLayout(self)
+        layout.setContentsMargins(0, 0, 0, 0)
 
         self.tab_widget = QTabWidget()
-        layout.addWidget(arg__1=self.tab_widget)
+        layout.addWidget(self.tab_widget)
 
         # save all tabs
         self.tabs: list[ModulTab] = []
 
         # hook tabs
         tab = DirectionsTab()
-        self.tab_widget.addTab(widget=tab, arg__2="Direction Data Extractor")
+        self.tab_widget.addTab(tab, "Direction Data Extractor")
         self.tabs.append(tab)
 
         tab = SpeedTab()
-        self.tab_widget.addTab(widget=tab, arg__2="Speed Data Extractor")
+        self.tab_widget.addTab(tab, "Speed Data Extractor")
         self.tabs.append(tab)
 
         tab = RoadObjectTab()
-        self.tab_widget.addTab(widget=tab, arg__2="Road Object Detector")
+        self.tab_widget.addTab(tab, "Road Object Detector")
         self.tabs.append(tab)
 
         tab = SegmentorTab()
-        self.tab_widget.addTab(widget=tab, arg__2="Road Segmentor")
+        self.tab_widget.addTab(tab, "Road Segmentor")
         self.tabs.append(tab)
 
         tab = PathTab()
-        self.tab_widget.addTab(widget=tab, arg__2="Path Planner")
+        self.tab_widget.addTab(tab, "Path Planner")
         self.tabs.append(tab)
 
         tab = FullAnnotationTab()
-        self.tab_widget.addTab(widget=tab, arg__2="Full Annotation Pipeline")
+        self.tab_widget.addTab(tab, "Full Annotation Pipeline")
         self.tabs.append(tab)
