--- tools/__init__.py
@@ -1,14 +1,0 @@
-from .basics import DirectionOutOfBounds, SpeedOutOfBounds
-from .types import Box, Extractors, Float, Img, ImgT, LImg, LImgT
-
-__all__: list[str] = [
-    "DirectionOutOfBounds",
-    "SpeedOutOfBounds",
-    "Float",
-    "Img",
-    "ImgT",
-    "LImg",
-    "LImgT",
-    "Box",
-    "Extractors",
-]
